Treat identity request responses as resolved promises
Instead of bailing out of of refetchDatum() for identity requests, this changes the logic to first create a resolved promise and then handle them the same as HTTP response to support composition and other post-response handlers. This simplifies the promise chain and fixes a potential bug with network error handling previously introduced with metadata support. This is also a step in the direction to provide for generic request handlers as this moves out some of the HTTP-specifics.