Inherit comparison during transformation
This sets the `comparison` to value of parent so that `comparison`s will passthrough with `then` and `catch` transformations. The transformations can still provide their own `comparison`, but that is unlikely to ever be needed because the parent request is what triggers the fetch and needs to be compared.

Fixes #53