Implement `connect.defaults()`
`connect.defaults()` returns a *new* `connect`, which means:

 - `.defaults()` calls can be chained.
 - The resulting function can be exported from a module, to provide
   cleaner APIs, and can be further modified closer to its application.
 - Defaults are local, not global to the application.

Defaults include both the options that may be set on fetch blocks and
functions in the `mapPropsToRequestsToProps`, and a number of internal
functions that can be used to hook into React Refetch's own processing
and replace or augment it with one's own.

Notably, the implementation of the `fetch` API can be controlled using
`.defaults()`. This feature also includes changes in the way `fetch` is
retrieved from the environment, checking both `window`, `global`, and
`self` (in this order), while previously only `window.fetch` was used.