Don't setTimeout if component is unmounted.
This happens in the case where the previous fetch hasn't returned yet
and the component is unmounted. In that case it would keep refreshing the
unmounted component forever.

Also includes fix for unrelated non-deterministic test that occurs because
the test assumes that two different lines in the test will take place more
than one millisecond apart, which is not always true.